Update Librería I2c

+++ Ejemplos/I2C/I2C-Master.X/I2C.h
@@ -17,6 +17,7 @@
 
 #include <xc.h> // include processor files - each processor file is guarded.  
 #include <pic16f887.h>
+#include <stdint.h>
 
 #ifndef _XTAL_FREQ
 #define _XTAL_FREQ 8000000
@@ -60,7 +61,7 @@ unsigned short I2C_Master_Read(unsigned short a);
 //*****************************************************************************
 // Función para inicializar I2C Esclavo
 //*****************************************************************************
-void I2C_Slave_Init(short address);
+void I2C_Slave_Init(uint8_t address);
 //*****************************************************************************
 #endif	/* __I2C_H */
