+++ Ejemplos/I2C/I2C-Slave.X/I2C.h
@@ -60,7 +60,7 @@ unsigned short I2C_Master_Read(unsigned short a);
 //*****************************************************************************
 // Función para inicializar I2C Esclavo
 //*****************************************************************************
-void I2C_Slave_Init(short address);
+void I2C_Slave_Init(uint8_t address);
 //*****************************************************************************
 #endif	/* __I2C_H */
